+++ app.js
@@ -63,7 +63,7 @@ app.use('/replies', require('./routes/replies'));
 app.use('/interview_experiences', require('./routes/interview_experiences'));
 app.use('/work_experiences', require('./routes/work_experiences'));
 app.use('/job_title_keywords', require('./routes/job_title_keywords'));
-app.use('/keywords', require('./routes/keywords'));
+app.use('/', require('./routes/company_keywords'));
 
 const corsOption = {
     origin: [
