+++ package.json
@@ -5,7 +5,6 @@
   "scripts": {
     "start": "node ./bin/www",
     "test": "istanbul cover ./node_modules/mocha/bin/_mocha --report lcovonly -- -R spec --recursive ",
-    "testonly": "mocha",
     "coveralls": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
     "lint": "eslint *.js libs/ routes/ test/ middlewares/ models/ database/",
     "lint:fix": "npm run lint -- --fix",
