+++ test/api/company_keywords/testKeywords.js
@@ -3,10 +3,10 @@ const request = require('supertest');
 const { MongoClient } = require('mongodb');
 const config = require('config');
 
-const app = require('../app');
-const create_capped_collection = require('../database/migrations/create-companyKeywords-collection');
+const app = require('../../../app');
+const create_capped_collection = require('../../../database/migrations/create-companyKeywords-collection');
 
-describe('keywords', () => {
+describe('company_keywords', () => {
     let db;
 
     before(() =>
@@ -27,31 +27,31 @@ describe('keywords', () => {
 
         it('will return keywords in order', () =>
             request(app)
-            .get('/keywords/company')
+            .get('/company_keywords')
             .query({ num: '2' })
             .expect(200)
             .expect((res) => {
-                assert.isArray(res.body);
-                assert.equal(res.body.length, 2);
-                assert.equal(res.body[0], 'GoodJob');
-                assert.equal(res.body[1], 'GoodJob2');
+                assert.isArray(res.body.keywords);
+                assert.equal(res.body.keywords.length, 2);
+                assert.equal(res.body.keywords[0], 'GoodJob');
+                assert.equal(res.body.keywords[1], 'GoodJob2');
             }));
 
         it('number should be 1~20', () =>
             request(app)
-            .get('/keywords/company')
+            .get('/company_keywords')
             .query({ num: '0' })
             .expect(422));
 
         it('number should be 1~20', () =>
             request(app)
-            .get('/keywords/company')
+            .get('/company_keywords')
             .query({ num: '100' })
             .expect(422));
 
         it('num should be integer number', () =>
             request(app)
-            .get('/keywords/company')
+            .get('/company_keywords')
             .query({ num: '10.5' })
             .expect(422));
